pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prueba-rag"
version = "0.1.0"
description = "RAG local sobre documentacion propia con Ollama remoto en LAN"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Equipo RAG" }
]
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.1",
    "typer>=0.12.3",
    "rich>=13.7.1",
    "faiss-cpu>=1.8.0",
    "numpy>=1.26.4",
    "unstructured>=0.14.5",
    "pypdf>=4.2.0",
    "markdown>=3.6",
    "tqdm>=4.66.4",
    "orjson>=3.10.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[project.scripts]
rag = "rag_local.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
